lowering: remove CallingConv variants, we only do SysV AMD64
