+++ lowering/src/lib.rs
@@ -68,9 +68,6 @@ pub fn run_backend(
     Ok(())
 }
 
-pub(crate) use crate::register::Amd64Reg;
-use std::convert::TryFrom;
-
 pub(crate) type VarId = (usize);
 
 /// FIXME refactor
@@ -79,38 +76,5 @@ fn var_id(op: lir::Operand) -> VarId {
     match op {
         Var(var) => (var.num()),
         Imm(_) => unreachable!(),
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub(crate) enum CallingConv {
-    X86_64,
-}
-
-impl CallingConv {
-    pub(self) fn num_arg_regs(self) -> usize {
-        match self {
-            CallingConv::X86_64 => 6,
-        }
-    }
-}
-
-#[derive(Debug, Hash, PartialEq, Eq, Copy, Clone)]
-pub(crate) enum Size {
-    One,
-    Four,
-    Eight,
-}
-
-impl TryFrom<u32> for Size {
-    type Error = String;
-
-    fn try_from(size: u32) -> Result<Self, Self::Error> {
-        match size {
-            1 => Ok(Size::One),
-            4 => Ok(Size::Four),
-            8 => Ok(Size::Eight),
-            x => Err(format!("only sizes 1,4 and 8 are supported, got {:?}", x)),
-        }
     }
 }
