+++ lowering/src/register.rs
@@ -1,5 +1,34 @@
-use crate::{CallingConv, Size};
 use std::{cmp::Ordering, collections::BTreeMap, convert::TryFrom};
+
+#[derive(Debug, Hash, PartialEq, Eq, Copy, Clone)]
+pub(crate) enum Size {
+    One,
+    Four,
+    Eight,
+}
+
+impl TryFrom<u32> for Size {
+    type Error = String;
+
+    fn try_from(size: u32) -> Result<Self, Self::Error> {
+        match size {
+            1 => Ok(Size::One),
+            4 => Ok(Size::Four),
+            8 => Ok(Size::Eight),
+            x => Err(format!("only sizes 1,4 and 8 are supported, got {:?}", x)),
+        }
+    }
+}
+
+#[derive(Copy, Clone, Debug)]
+pub(crate) struct CallingConv;
+
+impl CallingConv {
+    // FIXME this should be private
+    pub(crate) const fn num_arg_regs(self) -> usize {
+        6
+    }
+}
 
 #[derive(Debug, Hash, PartialEq, Eq, Copy, Clone)]
 pub(crate) struct Reg {
@@ -9,8 +38,7 @@ pub(crate) struct Reg {
 
 impl std::fmt::Display for Reg {
     fn fmt(&self, fmt: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        use self::Amd64Reg::*;
-        use super::Size::*;
+        use self::{Amd64Reg::*, Size::*};
         let (prefix, suffix) = match (self.reg, self.size) {
             (name, One) => match name {
                 A | B | C | D => ("", "l"),
@@ -294,9 +322,7 @@ impl RegisterAllocator {
     /// otherwise None.
     pub(super) fn arg_in_reg(&self, idx: usize) -> Option<Amd64Reg> {
         if idx < self.cconv.num_arg_regs() {
-            match self.cconv {
-                CallingConv::X86_64 => Some(Amd64Reg::arg(idx)),
-            }
+            Some(Amd64Reg::arg(idx)) // FIXME  this is confusing
         } else {
             None
         }
@@ -321,7 +347,7 @@ mod test {
 
     #[test]
     fn register_allocation_works() {
-        let mut allocator = RegisterAllocator::new(2, CallingConv::X86_64);
+        let mut allocator = RegisterAllocator::new(2, CallingConv);
 
         assert_eq!(allocator.alloc_reg().unwrap(), Amd64Reg::D);
